requirements.txt
tensorflow
keras
numpy
flask
pillow
matplotlib
scikit-learn